% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_fm.R
\name{cov_estim_afm}
\alias{cov_estim_afm}
\title{Approximate Factor Model Covariance Estimation}
\usage{
cov_estim_afm(data, factors = NULL, zeromean_log = FALSE, resid_est_func, ...)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{factors}{a nxf matrix with factors.
Default value is NULL and the factor is equal to the cross-sectional average of all the variables in the data.}

\item{zeromean_log}{a logical, indicating whether the data matrix has zero means (TRUE) or not (FALSE).
Default value is FALSE.}

\item{resid_est_func}{a covariance estimation function, applied to the residuals covariance matrix.}

\item{...}{further arguments to be parsed to resid_est_func}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, depending on the estimation function type.
}
}
\description{
Computes the Approximate Factor Model (AFM) estimator of the covariance matrix.
}
\details{
The AFM covariance estimator is calculated with the following formula:
\deqn{\hat{\Sigma} = B\hat{\Sigma}_F B' + \hat{\Sigma_u},}
where \eqn{\hat{\Sigma}_F} is the sample covariance matrix of the common factors and
\eqn{\hat{\Sigma}_u} is the residuals covariance matrix, estimated with the user-sapplied estim_func.
}
\examples{
data(rets_m)
# using the Ledoit-Wolf nonlinear shrinkage estimator
sigma_afm <- cov_estim_afm(rets_m, resid_est_func = cov_estim_lwnl)[[1]]
# using the Ledoit-Wolf linear shrinkage estimator with shrinkage intensity 0.1
sigma_afm <- cov_estim_afm(rets_m, resid_est_func = cov_estim_lwone, shrink_int = 0.1)[[1]]

}
