% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_fm.R
\name{cov_estim_efm}
\alias{cov_estim_efm}
\title{Exact Factor Model Covariance Estimation}
\usage{
cov_estim_efm(data, factors = NULL, zeromean_log = FALSE)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{factors}{a nxf matrix with factors.
Default value is NULL and the factor is equal to the cross-sectional average of all the variables in the data.}

\item{zeromean_log}{a logical, indicating whether the data matrix has zero means (TRUE) or not (FALSE).
Default value is FALSE.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here NA.
}
}
\description{
Computes the Exact Factor Model (EFM) estimator of the covariance matrix.
}
\details{
The EFM covariance estimator is calculated with the following formula:
\deqn{\hat{\Sigma} = B\hat{\Sigma}_F B' + \hat{\Sigma_u},}
where \eqn{\hat{\Sigma}_F} is the sample covariance matrix of the common factors and
\eqn{\hat{\Sigma}_u} is the covariaance matrix of residuals, assumed to have zero correlation.
}
\examples{
data(rets_m)
sigma_efm <- cov_estim_efm(rets_m)[[1]]

}
