% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_evc.R
\name{cov_estim_evc_bp}
\alias{cov_estim_evc_bp}
\title{Eigenvalue Clipping Covariance Estimation (Bouchaud-Potters)}
\usage{
cov_estim_evc_bp(data, cut_edge)
}
\arguments{
\item{data}{an nxp data matrix}

\item{cut_edge}{a double, indicating the proportion for the applied eigenvalue clipping.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the user-supplied cut edge.
}
}
\description{
Computes the Eigenvalue Clipping (EVC) estimator of the covariance matrix with the Bouchaud-Potters (BP) technique.
}
\details{
The eigenvalue clipping covariance matrix estimator is computed with the following formula:
\deqn{\hat{\Sigma}=\Delta\hat{\Lambda}\Delta',}
where \eqn{\Delta} is the matrix with the sample eigenvectors of the data matrix and
\eqn{\hat{\Lambda}} is a diagonal matrix with the clipped sample eigenvalues.
The clipping procedure follows \insertCite{bouchaudpotters2009;textual}{covestim}.
In particular, the user-defined cutting edge \eqn{s} gives the proportion for the applied eigenvalue clipping
so that the \eqn{(1-s)\times p} largest eigenvalues are kept
and the remaining \eqn{s\times p} eigenvalues are substituted by their average.
}
\examples{
data(rets_m)
sigma_evc_bp <- cov_estim_evc_bp(rets_m, cut_edge = 0.3)[[1]]

}
\references{
\insertAllCited
}
