% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_evc.R
\name{cov_estim_evc_mp}
\alias{cov_estim_evc_mp}
\title{Eigenvalue Clipping Covariance Estimation (Marcenko-Pastur)}
\usage{
cov_estim_evc_mp(data)
}
\arguments{
\item{data}{an nxp data matrix}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here an NA.
}
}
\description{
Computes the Eigenvalue Clipping (EVC) estimator of the covariance matrix with the Marcenko-Pastur (MP) edge.
}
\details{
The eigenvalue clipping covariance matrix estimator is computed with the following formula:
\deqn{\hat{\Sigma}=\Delta\hat{\Lambda}\Delta',}
where \eqn{\Delta} is the matrix with the sample eigenvectors of the data matrix and
\eqn{\hat{\Lambda}} is a diagonal matrix with the "clipped" sample eigenvalues.
The clipping procedure follows \insertCite{laloux1999;textual}{covestim}.
In particular, when assuming i.i.d returns, the eigenvalues of the sample correlation matrix
are distributed according to a Marcenko-Pastur distribution \insertCite{marvcenko1967distribution}{covestim} with
\deqn{\lambda_{min, max}=(1\mp\sqrt{p/n})^2}
as the smallest and largest eigenvalues of a random correlation matrix.
Therefore, only eigenvalues which lie outside this interval can bring useful information.
In this eigenvalue clipping procedure the sample eigenvalues bigger that \eqn{\lambda_{max}} are kept and
the rest are substituted with their average as in \insertCite{bouchaudpotters2009;textual}{covestim}.
}
\examples{
data(rets_m)
sigma_evc_mp <- cov_estim_evc_mp(rets_m)[[1]]

}
\references{
\insertAllCited
}
