% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_ewma.R
\name{cov_estim_ewma}
\alias{cov_estim_ewma}
\title{EWMA Covariance Estimation}
\usage{
cov_estim_ewma(data, lambda = 0.97)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{lambda}{a double for the decay parameter \eqn{\lambda}.
Default is 0.97 - the standard for monthly returns according to \insertCite{riskmetrics1996;textual}{covestim}.
If the data consists of daily returns, lambda should be set to 0.94.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the decay parameter \eqn{\lambda}.
}
}
\description{
Computes the Exponential Moving Average (EWMA) estimator of the covariance matrix.
}
\details{
The EWMA estimator of the covariance matrix for an nxp data matrix X is computed with the following formula:
\deqn{\hat{\Sigma}_{t}= (1-\lambda)R_{t}R'_{t} + \lambda\hat{\Sigma}_{t-1},}
where \eqn{R_{t}} is the matrix of demeaned returns for the time period \eqn{t} and
\eqn{\hat{\Sigma}_{t-1}} is the EWMA covariance estimator for the period \eqn{t-1}.
}
\examples{
data(rets_m)
sigma_ewma <- cov_estim_ewma(rets_m)[[1]]

}
\references{
\insertAllCited
}
