% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_lw.R
\name{cov_estim_lwcc_sf}
\alias{cov_estim_lwcc_sf}
\title{Ledoit-Wolf Covariance Estimation (Linear Shrinkage) IV}
\usage{
cov_estim_lwcc_sf(data, shrink_int = NULL, zeromean_log = FALSE)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{shrink_int}{a double, indicating the shrinkage intensity.
Default is the optimal shrinkage intensity as in \insertCite{ledoit2003identity;textual}{covestim}.}

\item{zeromean_log}{a logical, indicating whether the data matrix has zero means (TRUE) or not (FALSE).
Default value is FALSE.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the shrinkage intensity.
}
}
\description{
Computes the Ledoit-Wolf linear shrinkage estimator of the covariance matrix towards
the one-factor covariance matrix.
}
\details{
The Ledoit-Wolf linear shrinkage estimator of the covariance matrix towards
the one-factor covariance matrix is calculated with the following formula:
\deqn{\hat{\Sigma}= s\Sigma_{T} + (1-s)\Sigma,}
where \eqn{\Sigma} is the sample covariance matrix,
s is the user-supplied or optimal shrinkage intensity and
\eqn{\Sigma_{T}} is the covariance matrix estimator, given by a one-factor model,
where the factor is equal to the cross-sectional average of all the variables.
This covariance estimator assumes a zero correlation and variances of one
as the underlying covariance structure of the data.
A corresponding MATLAB code for the estimator can be accessed under
\url{https://www.econ.uzh.ch/en/people/faculty/wolf/publications.html}.
}
\examples{
data(rets_m)
sigma_lwcc_sf <- cov_estim_lwcc_sf(rets_m)[[1]]

}
\references{
\insertAllCited
}
