% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_lw.R
\name{cov_estim_lwone}
\alias{cov_estim_lwone}
\title{Ledoit-Wolf Covariance Estimation (Linear Shrinkage) II}
\usage{
cov_estim_lwone(data, shrink_int = NULL, zeromean_log = FALSE)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{shrink_int}{a double, indicating the shrinkage intensity.
Default is the optimal shrinkage intensity as in \insertCite{ledoit2004oneparam;textual}{covestim}.}

\item{zeromean_log}{a logical, indicating whether the data matrix has zero means (TRUE) or not (FALSE).
Default value is FALSE.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the shrinkage intensity.
}
}
\description{
Computes the Ledoit-Wolf linear shrinkage estimator of the covariance matrix towards the one-parameter matrix.
}
\details{
The Ledoit-Wolf linear shrinkage estimator of the covariance matrix towards
the diagonal matrix of equal variances is calculated with the following formula:
\deqn{\hat{\Sigma}= s\Sigma_{T} + (1-s)\Sigma,}
where \eqn{\Sigma} is the sample covariance matrix, s is the user-supplied or optimal shrinkage intensity and
\eqn{\Sigma_{T}} is a diagonal matrix with the average sample variance \eqn{\bar{\sigma}^2} on the diagonal.
This covariance estimator assumes a zero correlation and equal variances
as the underlying covariance structure of the data.
A corresponding MATLAB code for the estimator can be accessed under
\url{https://www.econ.uzh.ch/en/people/faculty/wolf/publications.html}.
}
\examples{
data(sp200)
sp_rets <- sp200[, -1]
sigma_lwone <- cov_estim_lwone(sp_rets)[[1]]

}
\references{
\insertAllCited
}
