% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_pca.R
\name{cov_estim_pca}
\alias{cov_estim_pca}
\title{Principal Component Analysis Covariance Estimation}
\usage{
cov_estim_pca(data, number_pc = NULL)
}
\arguments{
\item{data}{an nxp data matrix}

\item{number_pc}{an integer, indicating the number of principal components. Default value is NULL and
the number of principal components is set according to the Marcenko-Pastur edge
as in \insertCite{marvcenko1967distribution;textual}{covestim}.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the number of principal components.
}
}
\description{
Computes a Principal Component Analysis (PCA) estimator of the covariance matrix.
}
\examples{
data(sp200)
sp_rets <- sp200[1:100, -1]
# user-defined number of factors
sigma_pca <- cov_estim_pca(sp_rets, number_pc = 2)[[1]]
# number of factors, defined with MP cut-edge
results_pca_mp <- cov_estim_pca(sp_rets)
sigma_pca_mp <- results_pca_mp[[1]]
number_pc <- results_pca_mp[[2]]

}
\references{
\insertRef{johnson2002applied}{covestim}

\insertRef{fan2016overview}{covestim}
}
