% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_poet.R
\name{cov_estim_poet}
\alias{cov_estim_poet}
\title{Principal Orthogonal ComplEment Thresholding (POET) Covariance Estimation}
\usage{
cov_estim_poet(data, K = NULL, C = 1, thres = "soft", thres_mat = "vad")
}
\arguments{
\item{data}{an nxp data matrix.}

\item{K}{an integer, the number of principal components (factors).
Default values is NULL and the optimal K is calculated as in \insertCite{poetpackage;textual}{covestim}.
K=0 corresponds to threshoding the sample covariance directly.}

\item{C}{a double, the positive constant for thresholding. Default value is 1.}

\item{thres}{a character, indicating the choice of thresholding.
Possible values are "soft" for soft-thresholding, "hard" for hard thresholding and
"scad" for scad thresholding. Default is "soft".}

\item{thres_mat}{a character, indicating the option of thresholding either correlation or covairance matrix.
Possible values are "cor" for thresholding the error correlation matrix
then transform back to covariance matrix and
"vad" for thresholding the error covariance matrix directly.
Default is "cor".}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the number of principal components K.
}
}
\description{
Computes the POET estimator of the covariance matrix.
}
\details{
The POET estimator of the covariance matrix is computed
according to \insertCite{fan2013poet;textual}{covestim}.
The POET estimation is originally found under \insertCite{poetpackage;textual}{covestim}.
}
\examples{
data(sp200)
sp_rets <- sp200[, -1]
# user-defined K
sigma_poet <- cov_estim_poet(sp_rets, K = 2)[[1]]
# optimal K
sigma_poet <- cov_estim_poet(sp_rets)[[1]]

}
\references{
\insertAllCited
}
