% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_precond_sf.R
\name{cov_estim_precond_sf}
\alias{cov_estim_precond_sf}
\title{Covariance Estimation after Preconditioning with a Single Factor Model}
\usage{
cov_estim_precond_sf(data, zeromean_log = FALSE, precond_est_func = NULL, ...)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{zeromean_log}{a logical, indicating whether the data matrix has zero means (TRUE) or not (FALSE).
Default value is FALSE.}

\item{precond_est_func}{a function for estimating the precondtioned covariance matrix.}

\item{...}{further arguments to be parsed to precond_est_func}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the bandwidth speed.
}
}
\description{
Computes a specific estimator of the covariance matrix after preconditioning
the data with a Single Factor (SF) model - the implicit market portfolio.
}
\examples{
data(rets_m)
sigma_lwnl_sf <- cov_estim_precond_sf(rets_m,
  precond_est_func = cov_estim_lwnl, bandwidth_speed = NULL
)[[1]]
sigma_glasso_sf <- cov_estim_precond_sf(rets_m,
  precond_est_func = cov_estim_glasso, rho = 0.01
)[[1]]

}
\references{
\insertAllCited
}
