% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_ridge.R
\name{cov_estim_ridge}
\alias{cov_estim_ridge}
\title{Ridge-Penalized Covariance Estimation}
\usage{
cov_estim_ridge(data, rho)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{rho}{a double, indicating the ridge penalty.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the user-supplied ridge penalty \eqn{\rho}.
}
}
\description{
Computes the covariance matrix estimator with a ridge-penalty.
}
\details{
The ridge-penalized covariance matrix for a data matrix X is computed with the following formula:
\deqn{\hat{\Sigma}=V\left(\rho C+(1-\rho)I\right)V,}
where V is the sample volatility matrix, I is a pxp identity matrix,
C is the sample correlation matrix and \eqn{\rho} is the user-sapplied ridge penalty parameter.
}
\examples{
data(rets_m)
sigma_ridge <- cov_estim_ridge(rets_m, rho = 0.01)[[1]]

}
\references{
\insertRef{warton2008penalized}{covestim}
}
