% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_sh.R
\name{cov_estim_sh}
\alias{cov_estim_sh}
\title{Stein-Haff Covariance Estimation}
\usage{
cov_estim_sh(data)
}
\arguments{
\item{data}{an nxp data matrix.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here an NA.
}
}
\description{
Computes the Stein-Haff (SH) estimator of the covariance matrix.
}
\details{
The estimation procedure follows \insertCite{stein1975estimation;textual}{covestim} and
\insertCite{haff1991variational;textual}{covestim}.
Originally found under \insertCite{stcovpackage;textual}{covestim}.
}
\examples{
data(rets_m)
sigma_sh <- cov_estim_sh(rets_m)[[1]]

}
\references{
\insertAllCited
}
