% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_estim_tlasso.R
\name{cov_estim_tlasso}
\alias{cov_estim_tlasso}
\title{t-Lasso Covariance Estimation}
\usage{
cov_estim_tlasso(
  data,
  rho = NULL,
  pendiag_log = FALSE,
  df = 3,
  tol = 1e-05,
  maxit = 10000,
  symmetric_log = TRUE,
  theta_init = NULL
)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{rho}{a double, the non-negative regularization parameter \eqn{\rho} for lasso.
\eqn{\rho=0} means no regularization. Can be a scalar (usual) or a symmetric p by p matrix, or a vector of length p.
In the latter case, the penalty matrix has jkth element \eqn{\sqrt{\rho_j*\rho_k}}.
Default value is 0.}

\item{pendiag_log}{a logical, indicating whether the diagonal of the sample covariance matrix
is to be penalized (TRUE) or not (FALSE). Default value is FALSE.}

\item{df}{an integer, indicating the degrees of freedom of the assumed t-distribution. Default value is 3.}

\item{tol}{a double, indicating the tolerance for the glasso algorithm. Default value is set to 1e-05.}

\item{maxit}{an integer, indicating the maximum number of iterations for the glasso algorithm.
Default value is set to 10000.}

\item{symmetric_log}{a logical, indicating whether the output should be
a symmetric matrix (TRUE) or not necessarily (FALSE). Default value is set to TRUE.}

\item{theta_init}{a pxp initial matrix for the inverse of the covariance matrix.
Default value is NULL and the sample inverse for the t-distribution is used.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the lasso penalty.
}
}
\description{
Computes the t-Lasso (TLASSO) estimator of the covariance matrix.
}
\details{
The TLASSO estimator is elaborated in detail in \insertCite{finegold2011robust;textual}{covestim}.
Originally developed by \insertCite{torri2019sparse;textual}{covestim}.
}
\examples{
\dontrun{
data(sp200)
sp_rets <- sp200[, -1]
sigma_tlasso <- cov_estim_tlasso(sp_rets, rho = 0.001)[[1]]
}

}
\references{
\insertAllCited
}
