% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{is_posdef}
\alias{is_posdef}
\title{Matrix Positive-Definiteness Check}
\usage{
is_posdef(mat, tol = 1e-08)
}
\arguments{
\item{mat}{a pxp matrix.}

\item{tol}{a double with tolerance for the eigenvalues. Default value is 1e-8.}
}
\value{
a logical. TRUE, if the matrix mat is positive-definite.
}
\description{
Performs a tolerance check for positive-definiteness of a symmetric matrix.
}
\examples{
data(sp200)
sp_rets <- sp200[, -1]
ml_sigma <- cov_estim_wrapper2(sp_rets, "ML")
is_posdef(ml_sigma)

ml_sigma <- cov_estim_wrapper2(sp_rets[1:100, ], "ML")
is_posdef(ml_sigma)

}
