% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_help_funcs.R
\name{is_sparse}
\alias{is_sparse}
\title{Matrix Sparsity Check}
\usage{
is_sparse(mat)
}
\arguments{
\item{mat}{an nxp data matrix.}
}
\value{
a logical, indicating whether the matrix is sparse (TRUE) or not (FALSE).
}
\description{
Performs a sparsity check on a data matrix.
}
\examples{
data(rets_m)
sparse_check <- is_sparse(rets_m)

}
