% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{near_posdef}
\alias{near_posdef}
\title{Nearest Positive-Definite Matrix}
\usage{
near_posdef(
  mat,
  corr = FALSE,
  keep_diag = FALSE,
  do_dykstra = TRUE,
  eig_tol = 1e-06,
  conv_tol = 1e-07,
  posd_tol = 1e-08,
  maxit = 100,
  trace = FALSE
)
}
\arguments{
\item{mat}{a pxp matrix.}

\item{corr}{a logical, indicating if the result should be a correlation matrix. Default value is FALSE.}

\item{keep_diag}{a logical, indicating if the result should have the same diagonal as the original matrix mat.}

\item{do_dykstra}{a logical, indicating if Dykstra's correlation is to be used. Default value is TRUE.}

\item{eig_tol}{a double, defining the relative positiveness of eigenvalues
compared to the largest eigenvalue. Default value is 1e-6.}

\item{conv_tol}{a double, defining the convergence tolerance for the Higham algorithm. Default value is 1e-7.}

\item{posd_tol}{a double, defining the tolerance for enforcing positive definiteness. Default value is 1e-8.}

\item{maxit}{maximum number of iterations. Default value is 100.}

\item{trace}{a logical, indicating whether iterations are to be traced (printed out). Default value is FALSE.}
}
\value{
a positive-definite matrix.
}
\description{
Implements the algorithm of \insertCite{higham2002computing;textual}{covestim}
to compute the nearest positive-definite matrix to an approximate one,
typically a correlation or variance-covariance matrix.
}
\details{
Note that setting corr = TRUE just sets diag(.) <- 1 within the algorithm.
The near_posdef() is originally found under \insertCite{matrixpackage;textual}{covestim}.
}
\examples{

data(sp200)
sp_rets <- sp200[1:100, -1]
ml_sigma <- cov_estim_wrapper2(sp_rets, "ML")
ml_sigma_near_posdef <- near_posdef(ml_sigma, eig_tol = 1e-8)

}
\references{
\insertAllCited
}
