% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rets_m.R
\docType{data}
\name{rets_m}
\alias{rets_m}
\title{Monthly stock returns data from the S&P500 index}
\format{
An object of class \code{data.frame} with 216 rows and 200 columns.
}
\usage{
data(rets_m)
}
\description{
The dataset contains the monthly returns of 200 sampled stocks from the S&P500 index
for the time period from 12.2000 to 12.2018.
}
\keyword{datasets}
