% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-estim.R
\name{sigma_estim}
\alias{sigma_estim}
\title{Wrapper Function for Covariance Estimation II}
\usage{
sigma_estim(
  data,
  est_type,
  param = NULL,
  factors = NULL,
  zeromean_log = FALSE,
  res_all = FALSE
)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{est_type}{a character string, defining the estimation method.}

\item{param}{a double, setting an estimation specific (tuning) parameter.}

\item{factors}{a nxm data matrix of factors.}

\item{zeromean_log}{a logical, indicating whether the data matrix has zero means (TRUE) or not (FALSE). Default value is FALSE.}

\item{res_all}{a logical, defining the return object.
If FALSE, only the estimated covariance matrix is returned.
If TRUE, a list with two entries is returned. The first entry is the estimated covariance matrix.
The second entry is the estimation specific (tuning) parameter. Default value is FALSE.}
}
\value{
a list with following entries
\itemize{
\item a pxp estimated covariance matrix
\item an estimation specific tuning parameter
}
}
\description{
Estimates the covariance matrix of a dataset
according to the user-defined character string.
}
\examples{
data(sp200)
sp_rets <- sp200[,-1]
sigma_ml <- sigma_estim(sp_rets, "ML")[[1]]
sigma_lwcc <- sigma_estim(sp_rets, "LW-CC", param=0.3, res_all=TRUE)[[1]]

}
