% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-estim-simple.R
\name{sigma_estim_bs}
\alias{sigma_estim_bs}
\title{Bayes-Stein Covariance Estimation}
\usage{
sigma_estim_bs(data)
}
\arguments{
\item{data}{an nxp data matrix.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here an NA.
}
}
\description{
Computes the Bayes-Stein estimator of the covariance matrix.
}
\details{
The Bayes-Stein estimator of the covariance matrix is computed according to \insertCite{jorion1986bayes;textual}{CovEstim}.
}
\examples{
data(sp200)
sp_rets <- sp200[,-1]
sigma_bs <- sigma_estim_bs(sp_rets)[[1]]

}
\references{
\insertAllCited
}
