% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-estim-gh.R
\name{sigma_estim_ghs}
\alias{sigma_estim_ghs}
\title{Graphical Horseshoe Covariance Estimation}
\usage{
sigma_estim_ghs(data, burnin = 100, nmc = 5000)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{burnin}{an integer, indicating the number of MCMC burnins.}

\item{nmc}{an integer, indicating the number of samples.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here NA.
}
}
\description{
Computes the Graphical Horseshoe estimator of the covariance matrix.
}
\details{
The Graphical Horseshoe estimator is elaborated in detail in \insertCite{li2019graphical;textual}{CovEstim}.
More information on the functionality can be found in the Matlab package \url{https://github.com/liyf1988/GHS}.
}
\examples{
data(sp200)
sp_rets <- sp200[,-1]
sigma_ghs <- sigma_estim_ghs(sp_rets)[[1]]

}
\references{
\insertAllCited
}
