% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-estim-glasso.R
\name{sigma_estim_glasso}
\alias{sigma_estim_glasso}
\title{Graphical Lasso (GLASSO) Covariance Estimation}
\usage{
sigma_estim_glasso(
  data,
  rho = NULL,
  rho_seq = NULL,
  rho_seq_len = 10,
  rho_seq_min_ratio = 0.001,
  nfolds = 5,
  crit = "BIC",
  pendiag_log = FALSE,
  start = "cold",
  tol = 1e-05,
  maxit = 10000,
  cores = 1
)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{rho}{a double, the non-negative regularization parameter \eqn{\rho} for lasso. \eqn{\rho=0} means no regularization.
Can be a scalar (usual) or a symmetric p by p matrix, or a vector of length p. In the latter case, the penalty matrix has jkth element \eqn{\sqrt{\rho_j*\rho_k}}. Default value is NULL
and an optimal regularization is computed with the cross-validation (CV) procedure as in \insertCite{cvglassopackage;textual}{CovEstim}.}

\item{rho_seq}{a user-sapplied sequence of possible values for the regularization parameter \eqn{\rho}. Default value is NULL and
a sequence is defined as in \insertCite{cvglassopackage;textual}{CovEstim}.}

\item{rho_seq_len}{an integer, indicating the length for setting the rho sequence.}

\item{rho_seq_min_ratio}{a double, the minimum ratio for deviation between the rho values in the sequence.}

\item{nfolds}{an integer, indicating the number of folds for the CV. Default value is 5.}

\item{crit}{a character, indicating which selection criterion within the CV. Possible values are "loglik", "AIC" and "BIC". Default is set to "BIC".}

\item{pendiag_log}{a logical, indicating whether the diagonal of the sample covariance matrix is to be penalized (TRUE) or not (FALSE). Default value is FALSE.}

\item{start}{a character, specifying the start type of the glasso algorithm. Possible values are "warm" or "cold". Default value is "cold".}

\item{tol}{a double, indicating the tolerance for the glasso algorithm. Default value is set to 1e-05.}

\item{maxit}{an integer, indicating the maximum number of iterations for the glasso algorithm. Default value is set to 10000.}

\item{cores}{an integer, indicating how many cores should be used for the CV. Default value is 1. cores cannot be higher than the maximum number of cores of the processor in use.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the lasso penalty.
}
}
\description{
Computes the GLASSO estimator of the covariance matrix.
}
\details{
The GLASSO estimator is elaborated in detail in \insertCite{friedman2008sparse;textual}{CovEstim}. More information on the functionality can be found in
\insertCite{glassopackage;textual}{CovEstim} and \insertCite{cvglassopackage;textual}{CovEstim}.
}
\examples{
data(sp200)
sp_rets <- sp200[,-1]
sigma_glasso <- sigma_estim_glasso(sp_rets, rho=0.0001)[[1]]

}
\references{
\insertAllCited
}
