% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-estim-glasso.R
\name{sigma_estim_glasso_slim}
\alias{sigma_estim_glasso_slim}
\title{Graphical Lasso (GLASSO) Covariance Estimation - the slim (correlation) version}
\usage{
sigma_estim_glasso_slim(data, rho = NULL, type = "cor")
}
\arguments{
\item{data}{an nxp data matrix.}

\item{rho}{a double, the non-negative regularization parameter \eqn{\rho} for the lasso penalty.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the lasso penalty.
}
}
\description{
Computes the GLASSO estimator of the covariance matrix.
}
\details{
The GLASSO estimator is elaborated in detail in \insertCite{friedman2008sparse;textual}{CovEstim}. More information on the functionality can be found in
\insertCite{glassopackage;textual}{CovEstim} and \insertCite{cvglassopackage;textual}{CovEstim}.
}
\examples{
data(sp200)
sp_rets <- sp200[,-1]
sigma_glasso <- sigma_estim_glasso_slim(sp_rets, rho=0.01)[[1]]

}
\references{
\insertAllCited
}
