% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigma_estim_lwident_cpp}
\alias{sigma_estim_lwident_cpp}
\title{Ledoit-Wolf Linear Shrinkage Covariance Estimation I (CPP)}
\usage{
sigma_estim_lwident_cpp(data, shrink_int = -1, zeromean_log = FALSE)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{shrink_int}{a double, indicating the shrinkage intensity. Default is the optimal shrinkage intensity as in \insertCite{ledoit2003identity;textual}{CovEstim}.}

\item{zeromean_log}{a logical, indicating whether the data matrix has zero means (TRUE) or not (FALSE). Default value is FALSE.}
}
\value{
a pxp estimated covariance matrix.
}
\description{
Computes the Ledoit-Wolf linear shrinkage estimator of the covariance matrix towards the identity matrix.
}
\details{
The Ledoit-Wolf linear shrinkage estimator of the covariance matrix towards the identity matrix is calculated with the following formula:
\deqn{\hat{\Sigma}= s\Sigma_{T} + (1-s)\Sigma,}
where \eqn{\Sigma} is the sample covariance matrix, s is the user-supplied or optimal shrinkage intensity and \eqn{\Sigma_{T}} is a pxp identity matrix.
This covariance estimator assumes a zero correlation and variances of one as the underlying covariance structure of the data.
A corresponding MATLAB code for the estimator can be accessed under \url{https://www.econ.uzh.ch/en/people/faculty/wolf/publications.html}.
}
\examples{
data(sp200)
sp_rets <- sp200[,-1]
sigma_lwident <- sigma_estim_lwident_cpp(as.matrix(sp_rets))

}
\references{
\insertAllCited
}
