% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-estim-evc.R
\name{sigma_estim_pca}
\alias{sigma_estim_pca}
\title{PCA Covariance Estimation}
\usage{
sigma_estim_pca(data, number_pc = NULL)
}
\arguments{
\item{data}{an nxp data matrix}

\item{number_pc}{an integer, indicating the number of principal components. Default value is NULL and the
number of principal components is set according to the Marcenko-Pastur edge as in \insertCite{marvcenko1967distribution;textual}{CovEstim}.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the number of principal components.
}
}
\description{
Computes a PCA estimator of the covariance matrix.
}
\examples{
data(sp200)
sp_rets <- sp200[1:100,-1]
sigma_pca2 <- sigma_estim_pca(sp_rets, number_pc=2)[[1]] # user-defined number of factors
results_pca_mp <- sigma_estim_pca(sp_rets) # number of factors, defined with MP cut-edge
sigma_pca_mp <- results_pca_mp[[1]]
number_pc <- results_pca_mp[[2]]

}
\references{
\insertRef{johnson2002applied}{CovEstim}

\insertRef{fan2016overview}{CovEstim}
}
