% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-estim-factor.R
\name{sigma_estim_precond}
\alias{sigma_estim_precond}
\title{Covariance Estimation - Preconditioned}
\usage{
sigma_estim_precond(data, zeromean_log = FALSE, precond_est_func = NULL, ...)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{zeromean_log}{a logical, indicating whether the data matrix has zero means (TRUE) or not (FALSE). Default value is FALSE.}

\item{precond_est_func}{a function for estimating the precondtioned covariance matrix.}

\item{...}{further arguments to be parsed to precond_est_func.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here the bandwidth speed.
}
}
\description{
Computes a specific estimator of the covariance matrix after preconditioning
the data with a single factor model - the implicit market portfolio.
}
\examples{
data(sp200)
sp_rets <- sp200[,-1]
sigma_lwnl_sf <- sigma_estim_precond(sp_rets,
precond_est_func=sigma_estim_lwnl, bandwidth_speed=NULL)[[1]]
sigma_glasso_sf <- sigma_estim_precond(sp_rets,
precond_est_func=sigma_estim_glasso, rho=0.01)[[1]]

}
\references{
\insertAllCited
}
