% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-estim-evc.R
\name{sigma_estim_sh}
\alias{sigma_estim_sh}
\title{Stein-Haff Covariance Estimation}
\usage{
sigma_estim_sh(data)
}
\arguments{
\item{data}{an nxp data matrix.}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific tuning parameter, here an NA.
}
}
\description{
Computes the Stein-Haff estimator of the covariance matrix.
}
\details{
The estimation procedure follows \insertCite{stein1975estimation;textual}{CovEstim} and \insertCite{haff1991variational;textual}{CovEstim}.
Originally found under \insertCite{stcovpackage;textual}{CovEstim}.
}
\examples{
data(sp200)
sp_rets <- sp200[,-1]
sigma_sh <- sigma_estim_sh(sp_rets)[[1]]

}
\references{
\insertAllCited
}
