% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-estim.R
\name{sigma_estim_wrapper}
\alias{sigma_estim_wrapper}
\title{Wrapper Function for Covariance Estimation I}
\usage{
sigma_estim_wrapper(data, est_func, res_all = FALSE, ...)
}
\arguments{
\item{data}{an nxp data matrix.}

\item{est_func}{an estimation function.}

\item{res_all}{a logical, defining the return object.
If FALSE, only the estimated covariance matrix is returned.
If TRUE, a list with two entries is returned. The first entry is the estimated covariance matrix.
The second entry is the estimation specific (tuning) parameter. Default value is FALSE.}

\item{...}{additional parameters, parsed to est_func}
}
\value{
a list with the following entries
\itemize{
\item a pxp estimated covariance matrix.
\item an estimation specific (tuning) parameter.
}
}
\description{
Estimates the covariance matrix of a dataset
according to the user-defined function.
}
\examples{
data(sp200)
sp_rets <- sp200[,-1]
sigma_ml <- sigma_estim_wrapper(sp_rets, est_func=sigma_estim_ml)[[1]]


}
