% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{sigma_sim}
\alias{sigma_sim}
\title{Simulated Covariance Matrix}
\usage{
sigma_sim(p, corr_min, corr_max, vola_min, vola_max)
}
\arguments{
\item{p}{an integer, indicating the number of variables.}

\item{corr_min}{a double, indicating the minimum possible correlation across the p variables.}

\item{corr_max}{a double, indicating the maximum possible correlation across the p variables
(except the correlation with self of 1).}

\item{vola_min}{a double, indicating the minimum possible volatility of the p variables.}

\item{vola_max}{a double, indicating the maximum possible volatility of the p variables.}
}
\value{
a pxp covariance matrix.
}
\description{
Simulates a covariance matrix with specific properties
}
\details{
The correlation and volatility values are drown from a uniform distribution.
A condition for positive-definiteness of the resulting covariance matrix is not implemented.
}
\examples{
sim_sigma <- sigma_sim(100, corr_min = 0.01, corr_max = 0.40, vola_min = 0.05, vola_max = 0.30)

}
