% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{sqrt_root_mat_calc}
\alias{sqrt_root_mat_calc}
\title{Squared Root of a Matrix}
\usage{
sqrt_root_mat_calc(mat)
}
\arguments{
\item{mat}{a positive-definite pxp matrix.}
}
\value{
a pxp matrix, the squared root of mat.
}
\description{
Calculates the squared root of a matrix.
}
\details{
The squared root of a positive-definite symmetric matrix X is calculated with the following formula:
\deqn{\sqrt{X}=\Delta \sqrt{\Lambda}\Delta^{-1},}
where \eqn{\Delta} is the matrix of eigenvectors and
\eqn{\sqrt{\Lambda}} is a diagonal matrix with the squared roots of the eigenvalues of X.
}
\examples{
data(sp200)
sp_rets <- sp200[, -1]
sigma_ml <- cov_estim_wrapper2(sp_rets, "ML")
sigma_ml_sqrt_root <- sqrt_root_mat_calc(sigma_ml)

}
