% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_funcs.R
\name{is_sparse}
\alias{is_sparse}
\title{Matrix Sparsity Check}
\usage{
is_sparse(mat)
}
\arguments{
\item{mat}{an nxp data matrix.}
}
\value{
a logical, indicating whether the matrix is sparse (TRUE) or not (FALSE).
}
\description{
Performs a sparsity check on a data matrix.
}
\examples{
data(sp200)
sp_rets <- sp200[, -1]
sparsity_logic <- is_sparse(sp_rets)

}
